Continue the optional ffmpeg feature
