+++ examples/analyze.rs
@@ -1,4 +1,5 @@
-use bliss_audio::Song;
+use bliss_audio::decoder::ffmpeg::FFmpeg as Decoder;
+use bliss_audio::decoder::Decoder as DecoderTrait;
 use std::env;
 
 /**
@@ -9,7 +10,7 @@ use std::env;
 fn main() {
     let args: Vec<String> = env::args().skip(1).collect();
     for path in &args {
-        match Song::from_path(path) {
+        match Decoder::song_from_path(path) {
             Ok(song) => println!("{}: {:?}", path, song.analysis),
             Err(e) => println!("{path}: {e}"),
         }
