+++ src/chroma.rs
@@ -365,9 +365,13 @@ fn chroma_stft(
 mod test {
     use super::*;
     #[cfg(feature = "ffmpeg")]
+    use crate::song::decoder::ffmpeg::FFmpeg as Decoder;
+    #[cfg(feature = "ffmpeg")]
+    use crate::song::decoder::Decoder as DecoderTrait;
+    #[cfg(feature = "ffmpeg")]
     use crate::utils::stft;
     #[cfg(feature = "ffmpeg")]
-    use crate::{Song, SAMPLE_RATE};
+    use crate::SAMPLE_RATE;
     use ndarray::{arr1, arr2, Array2};
     use ndarray_npy::ReadNpyExt;
     use std::fs::File;
@@ -439,7 +443,7 @@ mod test {
     #[test]
     #[cfg(feature = "ffmpeg")]
     fn test_chroma_desc() {
-        let song = Song::decode(Path::new("data/s16_mono_22_5kHz.flac")).unwrap();
+        let song = Decoder::decode(Path::new("data/s16_mono_22_5kHz.flac")).unwrap();
         let mut chroma_desc = ChromaDesc::new(SAMPLE_RATE, 12);
         chroma_desc.do_(&song.sample_array).unwrap();
         let expected_values = vec![
@@ -462,7 +466,7 @@ mod test {
     #[test]
     #[cfg(feature = "ffmpeg")]
     fn test_chroma_stft_decode() {
-        let signal = Song::decode(Path::new("data/s16_mono_22_5kHz.flac"))
+        let signal = Decoder::decode(Path::new("data/s16_mono_22_5kHz.flac"))
             .unwrap()
             .sample_array;
         let mut stft = stft(&signal, 8192, 2205);
@@ -496,7 +500,7 @@ mod test {
     #[test]
     #[cfg(feature = "ffmpeg")]
     fn test_estimate_tuning_decode() {
-        let signal = Song::decode(Path::new("data/s16_mono_22_5kHz.flac"))
+        let signal = Decoder::decode(Path::new("data/s16_mono_22_5kHz.flac"))
             .unwrap()
             .sample_array;
         let stft = stft(&signal, 8192, 2205);
@@ -559,8 +563,10 @@ mod test {
 mod bench {
     extern crate test;
     use super::*;
+    use crate::song::decoder::ffmpeg::FFmpeg as Decoder;
+    use crate::song::decoder::Decoder as DecoderTrait;
     use crate::utils::stft;
-    use crate::{Song, SAMPLE_RATE};
+    use crate::SAMPLE_RATE;
     use ndarray::{arr2, Array1, Array2};
     use ndarray_npy::ReadNpyExt;
     use std::fs::File;
@@ -614,7 +620,7 @@ mod bench {
     #[bench]
     #[cfg(feature = "ffmpeg")]
     fn bench_chroma_desc(b: &mut Bencher) {
-        let song = Song::decode(Path::new("data/s16_mono_22_5kHz.flac")).unwrap();
+        let song = Decoder::decode(Path::new("data/s16_mono_22_5kHz.flac")).unwrap();
         let mut chroma_desc = ChromaDesc::new(SAMPLE_RATE, 12);
         let signal = song.sample_array;
         b.iter(|| {
@@ -626,7 +632,7 @@ mod bench {
     #[bench]
     #[cfg(feature = "ffmpeg")]
     fn bench_chroma_stft(b: &mut Bencher) {
-        let song = Song::decode(Path::new("data/s16_mono_22_5kHz.flac")).unwrap();
+        let song = Decoder::decode(Path::new("data/s16_mono_22_5kHz.flac")).unwrap();
         let mut chroma_desc = ChromaDesc::new(SAMPLE_RATE, 12);
         let signal = song.sample_array;
         b.iter(|| {
@@ -638,7 +644,7 @@ mod bench {
     #[bench]
     #[cfg(feature = "ffmpeg")]
     fn bench_chroma_stft_decode(b: &mut Bencher) {
-        let signal = Song::decode(Path::new("data/s16_mono_22_5kHz.flac"))
+        let signal = Decoder::decode(Path::new("data/s16_mono_22_5kHz.flac"))
             .unwrap()
             .sample_array;
         let mut stft = stft(&signal, 8192, 2205);
