+++ src/cue.rs
@@ -7,16 +7,13 @@
 //! and [CueInfo], which is a struct stored in [Song] to keep track of the CUE information
 //! the song was extracted from.
 
-#[cfg(feature = "ffmpeg")]
+use crate::song::decoder::Decoder as DecoderTrait;
 use crate::{Analysis, BlissError, BlissResult, Song, FEATURES_VERSION, SAMPLE_RATE};
 use rcue::cue::{Cue, Track};
-#[cfg(feature = "ffmpeg")]
 use rcue::parser::parse_from_file;
-#[cfg(feature = "ffmpeg")]
 use std::path::Path;
-use std::path::PathBuf;
-#[cfg(feature = "ffmpeg")]
 use std::time::Duration;
+use std::{marker::PhantomData, path::PathBuf};
 
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Default, Debug, PartialEq, Eq, Clone)]
@@ -34,12 +31,13 @@ pub struct CueInfo {
 }
 
 /// A struct to handle CUEs with bliss.
-/// Use either [analyze_paths](crate::analyze_paths) with CUE files or
-/// [songs_from_path](BlissCue::songs_from_path) to return a list of [Song]s
-/// from CUE files.
-pub struct BlissCue {
+/// Use either [analyze_paths](crate::decoder::Decoder::analyze_paths), which takes care of CUE files
+/// automatically, or [songs_from_path](BlissCue::songs_from_path) to return a list
+/// of [Song]s from CUE files.
+pub struct BlissCue<D: ?Sized> {
     cue: Cue,
     cue_path: PathBuf,
+    decoder: PhantomData<D>,
 }
 
 #[allow(missing_docs)]
@@ -54,16 +52,15 @@ struct BlissCueFile {
     audio_file_path: PathBuf,
 }
 
-impl BlissCue {
+impl<D: ?Sized + DecoderTrait> BlissCue<D> {
     /// Analyze songs from a CUE file, extracting individual [Song] objects
     /// for each individual song.
     ///
     /// Each returned [Song] has a populated [cue_info](Song::cue_info) object, that can be
     /// be used to retrieve which CUE sheet was used to extract it, as well
     /// as the corresponding audio file.
-    #[cfg(feature = "ffmpeg")]
     pub fn songs_from_path<P: AsRef<Path>>(path: P) -> BlissResult<Vec<BlissResult<Song>>> {
-        let cue = BlissCue::from_path(&path)?;
+        let cue: BlissCue<D> = BlissCue::from_path(&path)?;
         let cue_files = cue.files();
         let mut songs = Vec::new();
         for cue_file in cue_files.into_iter() {
@@ -84,7 +81,6 @@ impl BlissCue {
     }
 
     // Extract a BlissCue from a given path.
-    #[cfg(feature = "ffmpeg")]
     fn from_path<P: AsRef<Path>>(path: P) -> BlissResult<Self> {
         let cue = parse_from_file(&path.as_ref().to_string_lossy(), false).map_err(|e| {
             BlissError::DecodingError(format!(
@@ -93,14 +89,14 @@ impl BlissCue {
                 e
             ))
         })?;
-        Ok(BlissCue {
+        Ok(Self {
             cue,
             cue_path: path.as_ref().to_owned(),
+            decoder: PhantomData,
         })
     }
 
     // List all BlissCueFile from a BlissCue.
-    #[cfg(feature = "ffmpeg")]
     fn files(&self) -> Vec<BlissResult<BlissCueFile>> {
         let mut cue_files = Vec::new();
         for cue_file in self.cue.files.iter() {
@@ -114,7 +110,7 @@ impl BlissCue {
                 .iter()
                 .find(|(c, _)| c == "GENRE")
                 .map(|(_, v)| v.to_owned());
-            let raw_song = Song::decode(Path::new(&audio_file_path));
+            let raw_song = D::decode(Path::new(&audio_file_path));
             if let Ok(song) = raw_song {
                 let bliss_cue_file = BlissCueFile {
                     sample_array: song.sample_array,
@@ -134,7 +130,6 @@ impl BlissCue {
     }
 }
 
-#[cfg(feature = "ffmpeg")]
 impl BlissCueFile {
     fn create_song(
         &self,
@@ -211,12 +206,14 @@ mod tests {
     #[cfg(feature = "ffmpeg")]
     use super::*;
     #[cfg(feature = "ffmpeg")]
+    use crate::decoder::ffmpeg::FFmpeg;
+    #[cfg(feature = "ffmpeg")]
     use pretty_assertions::assert_eq;
 
     #[test]
     #[cfg(feature = "ffmpeg")]
     fn test_empty_cue() {
-        let songs = BlissCue::songs_from_path("data/empty.cue").unwrap();
+        let songs = BlissCue::<FFmpeg>::songs_from_path("data/empty.cue").unwrap();
         let error = songs[0].to_owned().unwrap_err();
         assert_eq!(
             error,
@@ -227,7 +224,7 @@ mod tests {
     #[test]
     #[cfg(feature = "ffmpeg")]
     fn test_cue_analysis() {
-        let songs = BlissCue::songs_from_path("data/testcue.cue").unwrap();
+        let songs = BlissCue::<FFmpeg>::songs_from_path("data/testcue.cue").unwrap();
         let expected = vec![
             Ok(Song {
                 path: Path::new("data/testcue.cue/CUE_TRACK001").to_path_buf(),
