+++ src/temporal.rs
@@ -95,14 +95,18 @@ impl Normalize for BPMDesc {
 mod tests {
     use super::*;
     #[cfg(feature = "ffmpeg")]
-    use crate::{Song, SAMPLE_RATE};
+    use crate::song::decoder::ffmpeg::FFmpeg as Decoder;
+    #[cfg(feature = "ffmpeg")]
+    use crate::song::decoder::Decoder as DecoderTrait;
+    #[cfg(feature = "ffmpeg")]
+    use crate::SAMPLE_RATE;
     #[cfg(feature = "ffmpeg")]
     use std::path::Path;
 
     #[test]
     #[cfg(feature = "ffmpeg")]
     fn test_tempo_real() {
-        let song = Song::decode(Path::new("data/s16_mono_22_5kHz.flac")).unwrap();
+        let song = Decoder::decode(Path::new("data/s16_mono_22_5kHz.flac")).unwrap();
         let mut tempo_desc = BPMDesc::new(SAMPLE_RATE).unwrap();
         for chunk in song.sample_array.chunks_exact(BPMDesc::HOP_SIZE) {
             tempo_desc.do_(&chunk).unwrap();
